Change key type for update timestamp

+++ src/MuHub.Domain/Entities/MarketCoinsUpdateTimestamp.cs
@@ -3,7 +3,7 @@ using MuHub.Domain.Common.Entities;
 
 namespace MuHub.Domain.Entities;
 
-public class MarketCoinsUpdateTimestamp : BaseEntity<int>
+public class MarketCoinsUpdateTimestamp : BaseEntity<long>
 {
     public DateTimeOffset LastUpdated { get; set; }
 }
